Create repeating Elements 1
In computer science, a set is an abstract data type that can store certain values, without any particular order, and no repeated values(Wikipedia). {1, 2, 3} is an example of a set ,but {1, 2, 2} is not a set. Today you will learn how to use sets in java by solving this problem.

You are given pairs of strings. Two pairs (a, b) and (c, d) are identical if a = c and b = d. That also implies (a, b) is not same as (b, a). After taking each pair as input, you need to print number of unique pairs you currently have.